refactor: implement persistent embedded storage

+++ androidApp/src/main/java/com/blackmidori/expenses/android/ui/workspace/chargesmodel/calculation/Bill.kt
@@ -2,4 +2,4 @@ package com.blackmidori.expenses.android.ui.workspace.chargesmodel.calculation
 
 import com.blackmidori.expenses.models.Expense
 
-class Bill (val expense: Expense, val amount: Double)
+class Bill (val expense: Expense, val amount: Float)
